chatrank/python-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
newspaper3k==0.2.8
beautifulsoup4==4.12.2
lxml>=5.0.0
google-generativeai==0.3.1
scikit-learn>=1.4.0
numpy>=1.26.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
rapidfuzz==3.5.2

